Terminal: Better row, column counts for a window size

+++ desktop/terminal/decoder.cpp
@@ -30,7 +30,7 @@ Decoder::Result Decoder::parse(char c)
                         };
                     
                     default:
-                        return { Result::Rune, c, {} };
+                        return { Result::Rune, (uint32_t)c, {} };
                 }
                 break;
